requirements/base.txt
pytz==2022.6  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.2.0  # https://github.com/evansd/whitenoise
redis==4.3.4  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
uvicorn[standard]==0.19.0  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==4.0.8  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.51.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-social-auth==0.3.6
python3-openid==3.2.0
social-auth-app-django==5.0.0
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
